fix: rename floating inputs to inset
