+++ src/components/index.ts
@@ -41,16 +41,16 @@ export * from './RadioGroup/RadioGroup';
 export * from './ResponsiveProvider/ResponsiveProvider';
 export * from './SelectInput/SelectInput';
 export * from './SelectInputNative/SelectInputNative';
-export * from './SelectInputNativeFloating/SelectInputNativeFloating';
+export * from './SelectInputNativeInset/SelectInputNativeInset';
 export * from './Spinner/Spinner';
 export * from './TabPanels/TabPanels';
 export * from './Tabs/Tabs';
 export * from './TabsSlider/TabsSlider';
 export * from './Table/Table';
 export * from './TextInput/TextInput';
-export * from './TextInputFloating/TextInputFloating';
+export * from './TextInputInset/TextInputInset';
 export * from './TextareaInput/TextareaInput';
-export * from './TextareaInputFloating/TextareaInputFloating';
+export * from './TextareaInputInset/TextareaInputInset';
 export * from './TextLink/TextLink';
 export * from './ThemeProvider/ThemeProvider';
 export * from './TimePicker/TimePicker';
